feat: [진건오] 일부 카테고리로 이동시 닉네임 출력이 안되는 이슈 해결 [BUCKET-LIST12]
