+++ p4th/GeonoJin/bucket_list_frontend/src/views/member/SignInView.vue
@@ -37,7 +37,6 @@ export default {
 
                 this.$store.state.isAuthenticated = true
 
-                this.$cookies.set("user", res.data, 3600);
                 localStorage.setItem("userInfo", JSON.stringify(res.data))
 
                 this.$store.state.currentLoginUserCheck = true
